@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
+using System.Text.Json;
 
-using Newtonsoft.Json;
 
 namespace GTranslatorAPI
 {
@@ -43,10 +43,7 @@ namespace GTranslatorAPI
         /// <param name="settingsFilePath">path of the file settings</param>
         public static Settings CreateFromFile(string settingsFilePath)
         {
-            var settings = JsonConvert.DeserializeObject<Settings>(
-                File.ReadAllText(settingsFilePath));
-            if (settings == null)
-                throw new InvalidOperationException("failed to create settings");
+            var settings = JsonSerializer.Deserialize<Settings>(File.ReadAllText(settingsFilePath));
             return settings;
         }
     }
